Moving business logic from controller to service
